+++ src/extensions/index.ts
@@ -1,1 +1,2 @@
 export * from "./grid-builder/index.js";
+export * from "./weather/index.js";